Record keycodes on key matrix.
This fixes a problem where a keycode can be set on one keyboard layer
and released on another. This will result in the keycode getting stuck.
This still probably needs to be modified so that the hid layer does not
have to know what the physical layout of the keyboard is.