Added key debouncing.
The debouncing is implemented in hardware using one counter per key. It
turns out it only uses 80 LUTs to implement on the Atreus, which is
reasonable in this particular case. Due to added latency of the 4bit
debounce counter I have also increased the matrix scanning clock by 2x.
This all might require some additional tweaking but so far it seems to
remedy the doubletaps...